Found a second, newer version

+++ src/lifnetwork/LIFNeuron.java
@@ -5,7 +5,6 @@
 package lifnetwork;
 
 import java.util.ArrayList;
-import java.util.concurrent.ThreadLocalRandom;
 
 /**
  *
@@ -17,18 +16,14 @@ public class LIFNeuron {
     final private NeuronType type;
     final private int rm;
     final private int cm;
-    final private int tau;
-    final private int refractoryPeriod;
+    final private int refracConst;
     final private int reversePotential;
-    final private int restPotential = -65;
+    final private int restPotential = -70;
     final private int threshold = -50;
-    //state dependent
     private float V;
     private float currentIn = 0;
-    private boolean firing = false;
-    private float refractoryTime;
-    private float synaticDynamics = 0;
-    final private ArrayList<synapticEvent> eventList;
+    private int timeInFiring=Integer.MAX_VALUE;
+
     final private ArrayList<IncomingSynapse> incomingList;
 
     /**
@@ -36,19 +31,16 @@ public class LIFNeuron {
      * @param type
      * @param r
      * @param c
-     * @param tau
      * @param refractoryPeriod
      * @param reversePotential
      */
-    public LIFNeuron(NeuronType type, int r, int c, int tau, int refractoryPeriod, int reversePotential) {
+    public LIFNeuron(NeuronType type, int r, int c, int refractoryPeriod, int reversePotential) {
         this.type = type;
         this.rm = r;
         this.cm = c;
-        this.tau = tau;
-        this.refractoryPeriod = refractoryPeriod;
+        this.refracConst = refractoryPeriod;
         this.reversePotential = reversePotential;
 
-        this.eventList = new ArrayList<>();
         this.incomingList = new ArrayList<>();
         this.V = restPotential;
     }
@@ -61,19 +53,15 @@ public class LIFNeuron {
         this.currentIn += i;
     }
 
-    public boolean isFiring() {
-        return firing;
-    }
-
     public void updateCurrentInput() {
         currentIn = 0;
         for (IncomingSynapse incoming : incomingList) {
-            if (incoming.getSynapticDynamics() > 0) {
+//            if (incoming.getPre().getTimeInFiring() < refracConst) {
+            if (incoming.getPre().getTimeInFiring() < 50000) {
                 /*
                  * accumulate current inputs;
                  */
-                //Current(post)=Current(post)+(s(pre))*(ReversePotential(pre)-NeuronsV(post))*g(pre,post)*Weight(pre,post);
-                currentIn += incoming.getSynapticDynamics() * (incoming.getReversePotential() - this.V) * incoming.getG() * incoming.getWeight();
+                currentIn += SynapticEvent.getEventDynamics(incoming.getPre().getType(), this.type, incoming.getPre().getTimeInFiring()) * (incoming.getReversePotential() - this.V) * incoming.getDrivingForce();
             }
         }
     }
@@ -82,8 +70,12 @@ public class LIFNeuron {
         return this.reversePotential;
     }
 
-    public float getSynapticDynamics() {
-        return synaticDynamics;
+    public int getTimeInFiring() {
+        return timeInFiring;
+    }
+
+    public NeuronType getType() {
+        return type;
     }
 
     public float getV() {
@@ -94,72 +86,29 @@ public class LIFNeuron {
         return currentIn;
     }
 
-    public void updateSynapticDynamics(int currentTime) {
-
-        if (eventList.isEmpty()) {
-            return;
-        }
-        float max = 0;
-        boolean[] toRemove = new boolean[eventList.size()];
-        for (int i = 0; i < eventList.size(); i++) {
-//            System.out.print("-1,");
-            float eventDynamics = eventList.get(i).getEventDynamics(currentTime);
-//            System.out.print("1,");
-            if (eventDynamics == -1) {
-                toRemove[i] = true;
-            } else if (eventDynamics > max) {
-                max = eventDynamics;
-            }
-        }
-        synaticDynamics = max;
-        int removed = 0;
-        for (int i = 0; i < toRemove.length; i++) {
-            if (toRemove[i]) {
-                eventList.remove(i - removed);
-                removed++;
-            }
-        }
+    public void updateTimeInFiring(int dT) {
+        timeInFiring += dT;
     }
 
-    private final class synapticEvent {
-
-        final private float scale;
-        final private int riseTime = 2000;
-        final private int synaticDelay = 2000;
-        final private int eventTime;
-
-        public synapticEvent(int eventTime) {
-            this.eventTime = eventTime;
-            scale = ThreadLocalRandom.current().nextFloat();
-        }
-
-        public float getEventDynamics(int currentTime) {
-
-            int postAPTime = currentTime - eventTime;
-            if (postAPTime <= synaticDelay) {
-                return 0;
-            } else if ((postAPTime <= synaticDelay + riseTime)) {
-                return (float) (postAPTime - synaticDelay) / riseTime;
-            } else {
-                float value = (float) Math.pow(Math.E, (double) (synaticDelay + riseTime - postAPTime) / tau) * scale;
-//                float value = (synaticDelay + riseTime - postAPTime) + (2 * tau);
-                return value > 0.001 ? value : -1;
-            }
-        }
+    public void triggerFire(){
+        this.V=30;
     }
 
-    public boolean updateVoltageAndFire(int dT, int currentTime) {
-        this.V += (float) dT * (currentIn - (this.V - restPotential) / (rm / 1000f))
-                / this.cm / 1000f;//rm*1000,dT*1000
-        firing = (this.V > this.threshold) && (refractoryTime <= 0);
-        this.V = (this.V > this.threshold) ? restPotential : this.V;
-        if (firing) {
-            refractoryTime = refractoryPeriod;
-            this.eventList.add(new synapticEvent(currentTime));
-            return true;
-        } else {
-            refractoryTime -= refractoryTime > 0 ? dT : 0;
+    public boolean updateVoltageAndFire(int dT) {
+        if (timeInFiring < refracConst) {
+            this.V = restPotential;
+            timeInFiring += dT;
             return false;
+        } else {
+            this.V += (float) dT * (currentIn - (this.V - restPotential) / (rm / 1000f))
+                    / this.cm / 1000f;//rm*1000,dT*1000
+            if (this.V > this.threshold) {
+                timeInFiring = 0;
+                return true;
+            } else {
+                return false;
+            }
         }
     }
+
 }
